+++ android/src/main/kotlin/io/flutter/plugins/geofencing/GeofencingBroadcastReceiver.kt
@@ -7,16 +7,17 @@ package io.flutter.plugins.geofencing
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import io.flutter.view.FlutterMain
+import io.flutter.embedding.engine.loader.FlutterLoader
 
 
 class GeofencingBroadcastReceiver : BroadcastReceiver() {
     companion object {
         private const val TAG = "GeofencingBroadcastReceiver"
     }
+
     override fun onReceive(context: Context, intent: Intent) {
-        FlutterMain.startInitialization(context)
-        FlutterMain.ensureInitializationComplete(context, null)
+        FlutterLoader().startInitialization(context)
+        FlutterLoader().ensureInitializationComplete(context, null)
         GeofencingService.enqueueWork(context, intent)
     }
 }