+++ android/src/main/kotlin/io/flutter/plugins/geofencing/IsolateHolderService.kt
@@ -12,7 +12,6 @@ import android.content.Intent
 import android.os.IBinder
 import android.os.PowerManager
 import androidx.core.app.NotificationCompat
-import android.util.Log
 import io.flutter.embedding.engine.FlutterEngine
 
 class IsolateHolderService : Service() {
@@ -42,7 +41,7 @@ class IsolateHolderService : Service() {
         val channel = NotificationChannel(CHANNEL_ID,
                 "Flutter Geofencing Plugin",
                 NotificationManager.IMPORTANCE_LOW)
-        val imageId = getResources().getIdentifier("ic_launcher", "mipmap", getPackageName())
+        val imageId = resources.getIdentifier("ic_launcher", "mipmap", packageName)
 
         (getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager).createNotificationChannel(channel)
         val notification = NotificationCompat.Builder(this, CHANNEL_ID)
@@ -62,10 +61,10 @@ class IsolateHolderService : Service() {
     }
 
     override fun onStartCommand(intent: Intent, flags: Int, startId: Int): Int {
-        if (intent.getAction() == ACTION_SHUTDOWN) {
+        if (intent.action == ACTION_SHUTDOWN) {
             (getSystemService(Context.POWER_SERVICE) as PowerManager).run {
                 newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, WAKELOCK_TAG).apply {
-                    if (isHeld()) {
+                    if (isHeld) {
                         release()
                     }
                 }